requirements.txt
biopython==1.68
ete2==2.3.10
gensim==0.13.2
h5py==2.7.1
Keras==2.0.8
keras-spp==0.0.0
Markdown==2.6.9
matplotlib==2.1.0
mock==2.0.0
ncbi-genome-download==0.2.5
networkx==2.1
numpy==1.11.1
olefile==0.44
pandas==0.22.0
pbr==3.1.1
pep8==1.7.0
protobuf==3.4.0
py==1.4.33
pycparser==2.18
pyOpenSSL==17.5.0
pyparsing==2.2.0
pytest==3.0.7
python-apt==1.1.0b1
python-dateutil==2.6.0
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
requests==2.13.0
s3transfer==0.1.12
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==0.19.0
seaborn==0.8.1
sh==1.12.14
six==1.10.0
smart-open==1.5.1
subprocess32==3.2.7
tensorboard==1.0.0a6
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.8
tox==2.7.0
tox-pyenv==1.0.3
tqdm==4.19.5
unity-lens-photos==1.0
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.12.2
